+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### [5.0.4](https://github.com/rizen/pixi-glyphs/compare/v5.0.3...v5.0.4) (2025-11-10)
 
+- Fix: Icon position is not offset by outline the way text is.
+- Fix: Icons are not getting effects the way text is.
+
+
 ### [5.0.3](https://github.com/rizen/pixi-glyphs/compare/v5.0.2...v5.0.3) (2025-11-06)
 
 - Fix: Inline icons were adding an inappropriate descender extension causing gaps between lines.